Fix warnings caused by 64-bit migration

+++ DLStarRating/DLStarRatingControl.m
@@ -62,7 +62,7 @@
 	self = [super initWithFrame:frame];
 	if (self) {
         isFractionalRatingEnabled = isFract;
-		numberOfStars = _numberOfStars;
+		numberOfStars = (int)_numberOfStars;
         if (isFractionalRatingEnabled)
             numberOfStars *=kNumberOfFractions;
 		[self setupView];
@@ -80,7 +80,7 @@
 #pragma mark Customization
 
 - (void)setStar:(UIImage*)defaultStarImage highlightedStar:(UIImage*)highlightedStarImage {
-  for(NSInteger i = 0; i < numberOfStars; i++){
+  for(int i = 0; i < numberOfStars; i++){
     [self setStar:defaultStarImage highlightedStar:highlightedStarImage atIndex:i];
   }
 }
@@ -136,7 +136,7 @@
 	CGPoint point = [touch locationInView:self];	
 	UIButton *pressedButton = [self starForPoint:point];
 	if (pressedButton) {
-		int idx = pressedButton.tag;
+		int idx = (int)pressedButton.tag;
 		if (pressedButton.highlighted) {
 			[self disableStarsDownToExclusive:idx];
 		} else {
@@ -156,7 +156,7 @@
 	
 	UIButton *pressedButton = [self starForPoint:point];
 	if (pressedButton) {
-		int idx = pressedButton.tag;
+		int idx = (int)pressedButton.tag;
 		UIButton *currentButton = (UIButton*)[self subViewWithTag:currentIdx];
 		
 		if (idx < currentIdx) {
